requirements/base.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt[blacklist]>=5.3
psycopg[binary]>=3.2
python-dotenv>=1.0
celery>=5.4
redis>=5.0
cryptography>=43.0
django-filter>=24.2
drf-spectacular>=0.27
django-storages>=1.14
boto3>=1.34